api/requirements.txt
anthropic==0.30.1
fastapi==0.99.0
mangum==0.17.0
pydantic
pymongo==4.7.3
PyPDF2==3.0.1
uvicorn==0.30.1
python-multipart
python-dotenv